Rename test.py -> evaluate_dqn.py
